+++ lib/devall_analytics.dart
@@ -1,5 +1,6 @@
 library devall_analytics;
 
 export 'src/devall_analytics_core.dart';
+export 'src/offline_storage.dart';
 export 'enums.dart';
 export 'device_identity.dart';